Added example of using custom built PDFs.
The special thing here is that I'm using CUDA 5.0 ability to link device code so
that I don't have to rebuild built-in PDFs. Well.... That's how it is supposed
to work in principle, but since I don't have access to the cluster I can't build
the central GooPdfCuda.o properly with the '-dc' flag. So I'm stuck rebuilding
it locally, in the folder. But hey, at least I don't have to append new PDFs to
CUDAglob.cu.